refactor: ajuste na estrutura do repositório
