+++ setup.cfg
@@ -1,0 +1,50 @@
+[metadata]
+name = togglReports
+author = Rodrigo Mendonça
+author_email = rodrigogr.mendonca@gmail.com
+description = A simple project to automate the creation of reports using the Toggl API .
+long_description = file:README.md
+url = https://github.com/ro-56/python_package_skeleton
+classifiers =
+	Development Status :: 1 - Planning
+	Intended Audience :: Developers
+	License :: OSI Approved :: MIT License
+	Programming Language :: Python :: 3
+	Programming Language :: Python :: 3 :: Only
+
+[options]
+packages = 
+	togglReports
+include_package_data = true
+python_requires = >=3.9
+package_dir = 
+	=src
+
+install_requires =
+    pandas >= 1.2.4
+    requests >= 2.25.*
+    datetime >= 2.7.*
+    pyyaml >= 6.0
+    argparse >= 1.3.*
+
+[options.packages.find]
+exclude =
+	build*
+	dist*
+	docs*
+	tests*
+
+[options.extras_require]
+testing =
+	pytest >= 6.2
+	pytest-cov >= 2.0
+	mypy >= 0.910
+	flake8 >= 3.9
+	tox >= 3.24
+
+[options.package_data]
+amslint = py.typed
+
+[options.entry_points]
+console_scripts =
+    getSguToggl=togglReports.cli.cli:main
